hello library and tested greeting program that uses it
The greeting program is part of the hello project because it's also
our test for that project (at least until we get a unit test framework
in place).

With this you can confirm that, even with enable_testing() in
hello/CMakeLists.txt, you still need to call enable_testing() at the
top level as well for `ctest` to work at the top level. When the top
level has it, it's not necessary to put it in the subproject's
hello/CMakeLists.txt, but we add it there anyway so that it can be
built and tested standalone.

The implementation of the greet() function sucks and is not thread-safe.