Test: On Windows always build for x64 architecture
This assumes the default "Visual Studio 15 2017" generator that CMake
uses when called from the command line. I'm not sure if the `-A x64`
platform parameter will also work with Ninja. (It definitely doesn't
with the Makefile generator, which is one reason why we don't set it
on Unix.)

This fixes the issue where FindPython3 couldn't find an interpreter on
Windows. (It's not clear how it found the development files; perhaps
it mistook the presumably 64-bit library for a 32-bit library.)
However, we still have the following two issues on Windows:
1. Without manual selection of a release configuration it builds a
   debug configuration by default and cannot find python36_d.lib. (The
   cache contains Python3_LIBRARY_DEBUG set to
   Python3_LIBRARY_DEBUG-NOTFOUND.)
2. With manual selection of a release configuration in Visual Studio,
   it still can't find python36.lib, even though in the cache
   Python3_LIBRARY_RELEASE is set to what appears to be the correct
   file on my system, C:/Program Files/Python36/libs/python36.lib.