hello-pyext/: Use Python3_add_library() and make it a MODULE
Python3_add_library() handles setting the include dir and giving the
generated Python extension the correct name (no `lib` prefix on Unix,
`.PYD` extension instead of `.DLL` on Windows), so we remove the hacks
we had to do both of those.

The library needs to be a MODULE, because that, not SHARED, is the
correct setting for shared libs that are loaded at runtime by a
program rather than linked automatically by the system linker. (SHARED
works anyway on Linux, but not Windows.)

The test still doesn't pass with MSBuild because that's a
multi-configuration build tool and so the generated file ends up in a
subdirectory named for the configuration. That needs to be fixed, but
I have verified via a manual copy that Python can load and use the
extension.