hello-pyext/ → pyext/ because better command-line completion
Having a common `hello` prefix on both subprojects made for a lot
more typing when using command-line completion, so fix that.

We leave the pyext/'s project name as `hello-pyext` because it feels
more clear and as an experiment to see what kind of problems or
infelicities this might cause.