pyext/: Fix method of finding target directory for PYTHONPATH
We need to add the directory into which our phello extension was built
to PYTHONPATH so that test.py can load it.

The output file for a target is in CMAKE_CURRENT_BINARY_DIR only with
single-configuration generators; for multi-configuration generators
such as MSBuild it will be in a subdirectory for the current
configuration we're building. So replace CMAKE_CURRENT_BINARY_DIR with
$<TARGET_FILE_DIR:phello>, which will include the subdir if that's
being used.